pyproject.toml
[tool.poetry]
name = "stockpredict"
version = "0.1.0"
description = "MLOps project for azure ml"
authors = ["Bogdan Sapovskyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.8"
pandas = "2.0.3"
deltalake = "0.10.1"
mlflow = "2.5.0"
azure-identity = "1.14.0"
azure-ai-ml = "1.9.0"
azureml-mlflow = "1.53.0"
mldesigner = "0.1.0b14"
azureml-core = "1.52.0"
azureml-dataset-runtime = "1.52.0"
azureml-defaults = "1.52.0"
azure-ml = "0.0.1"
azure-ml-component = "0.9.18.post2"
azureml-contrib-services = "1.52.0"
azureml-automl-common-tools = "1.52.0"
azureml-inference-server-http = ">=0.8.0,<0.9.0"
inference-schema = ">=1.5.0,<1.6.0"
joblib = "1.3.2"
azure-keyvault-secrets = "4.7.0"
aiohttp = "3.8.5"
lightning = "2.0.9"
nest-asyncio = "1.5.8"
numpy = "1.23.5"
torch = "2.0.1"



[tool.poetry.group.dev.dependencies]
pyspark = "3.3.0"
azure-mgmt-msi = "7.0.0"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
register_azure = 'stockpredict.deployment.main:main'